refactor : refactoring spring data jdbc code
