+++ src/main/java/com/example/SpringHelloApplication.java
@@ -1,4 +1,4 @@
-package com.example.springhello;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
